requirements.txt
streamlit
pandas
numpy
requests
openai
google-generativeai
folium
streamlit-folium
geopy
python-dotenv
openrouteservice